+++ backend/cli/tests/assets/cms-index-vendor-dir-files.php
@@ -6,6 +6,7 @@ return [
     "\$index/public/sivujetti/vendor/jspanel.min.js",
     "\$index/public/sivujetti/vendor/pickr-theme-nano.min.css",
     "\$index/public/sivujetti/vendor/pickr.min.js",
+    "\$index/public/sivujetti/vendor/popper.min.js",
     "\$index/public/sivujetti/vendor/vendor.bundle.min.css",
     "\$index/public/sivujetti/vendor/vendor.bundle.min.js",
 ];