+++ backend/sivujetti/tests/src/Block/RenderEachBuiltInBlockTest.php
@@ -298,17 +298,21 @@ final class RenderEachBuiltInBlockTest extends RenderBlocksTestCase {
                     "renderWith" => "sivujetti:block-listing-pages-default"],
                 id: "@auto"),
         ];
-        $state->makeExpectedHtml = fn(object $b, object $pageData) =>
-            "<div class=\"listing listing-pages\" data-block-type=\"Listing\" data-block=\"{$b->id}\">\r\n" .
-            "        <article class=\"list-item list-item-/hello\">\r\n" .
+        $makeExpectedListItem = fn(object $pageData) =>
+            "        <article class=\"list-item list-item-{$pageData->slug}\">\r\n" .
             "        <h2>".Template::e($pageData->title)."</h2>\r\n" .
             "        <div><a href=\"".Template::makeUrl($pageData->slug)."\">Read more</a></div>\r\n" .
-            "    </article>\r\n" .
+            "    </article>\r\n";
+        $state->makeExpectedHtml = fn(object $b, object $page1Data, object $page2Data = null) =>
+            "<div class=\"listing listing-pages\" data-block-type=\"Listing\" data-block=\"{$b->id}\">\r\n" .
+                ($makeExpectedListItem($page1Data)) .
+                ($page2Data ? $makeExpectedListItem($page2Data) : "") .
             "    </div>\r\n";
         return $state;
     }
     private function insertTestPages(\TestState $state): void {
-        $this->pageTestUtils->insertPage($state->testPageData, $state->customPageType ?? null);
+        foreach (!is_array($state->testPageData) ? [$state->testPageData] : $state->testPageData as $pageData)
+            $this->pageTestUtils->insertPage($pageData, $state->customPageType ?? null);
     }
     public function testRenderBlockRendersCustomListings(): void {
         $makeExpectedHtml = fn($pageCount, $firstPageTitle) =>
@@ -345,6 +349,48 @@ final class RenderEachBuiltInBlockTest extends RenderBlocksTestCase {
     }
     private function dropCustomPageType(\TestState $state): void {
         $this->pageTestUtils->dropCustomPageType($state->customPageType);
+    }
+    public function testListingBlockUsesLimitFilter(): void {
+        $state = $this->setupRenderListingBlockHavingManyListItemsTest();
+        $this->makeTestSivujettiApp($state);
+        $this->insertTestPages($state);
+        $fn = $state->makeExpectedHtml;
+        //
+        $expectedListItemsWithoutLimit = [$state->testPageData[1], $state->testPageData[0]];
+        $expectedHtml = $fn($state->testBlocks[0], ...$expectedListItemsWithoutLimit);
+        $this->renderAndVerify($state, 0, $expectedHtml);
+        //
+        $this->blockTestUtils->setBlockProp($state->testBlocks[0], "filterLimit", 1);
+        $expectedListItemsWithLimit = [$state->testPageData[1]];
+        $expectedHtml = $fn($state->testBlocks[0], ...$expectedListItemsWithLimit);
+        $this->renderAndVerify($state, 0, $expectedHtml);
+    }
+    private function setupRenderListingBlockHavingManyListItemsTest(): \TestState {
+        $state = $this->setupRenderListingBlocksTest();
+        $page2 = clone $state->testPageData;
+        $page2->slug = "/something-else";
+        $page2->path = "something-else/";
+        $state->testPageData = [$state->testPageData, $page2];
+        return  $state;
+    }
+    public function testListingBlockUsesOrderFilter(): void {
+        $state = $this->setupRenderListingBlockHavingManyListItemsTest();
+        $this->makeTestSivujettiApp($state);
+        $this->insertTestPages($state);
+        $fn = $state->makeExpectedHtml;
+        //
+        $expectedListItemsWithDescendingOrderFilter = [$state->testPageData[1], $state->testPageData[0]];
+        $expectedHtml = $fn($state->testBlocks[0], ...$expectedListItemsWithDescendingOrderFilter);
+        $this->renderAndVerify($state, 0, $expectedHtml);
+        //
+        $this->blockTestUtils->setBlockProp($state->testBlocks[0], "filterOrder", "asc");
+        $expectedListItemsWithAscendingOrderFilter = [$state->testPageData[0], $state->testPageData[1]];
+        $expectedHtml = $fn($state->testBlocks[0], ...$expectedListItemsWithAscendingOrderFilter);
+        $this->renderAndVerify($state, 0, $expectedHtml);
+        //
+        $this->blockTestUtils->setBlockProp($state->testBlocks[0], "filterOrder", "rand");
+        $this->sendRenderBlockRequest($state, $state->testBlocks[0]);
+        $this->verifyRequestFinishedSuccesfully($state);
     }