+++ changelog.md
@@ -6,12 +6,14 @@
 
 ## Added
 - An ability to drag blocks between block tree's branches
+- `filterLimit` to Listing blocks
 
 ## Fixed
 - An issue, where the page blew up in edit mode, if some of the blocks contained some unusual css
 
 ## Changed
 - The event type (onChange -> onInput) that refreshes styles preview in edit mode
+- The UI of Listing blocks
 
 # 0.9.0 (2022-04-08)